+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.25.0] - 2026-07-02
+
 ### Added
 
 - **The wiki is now searchable prose.** Accepted wiki pages are split into sections